requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
ansible-base==2.10.17
    # via ansible
ansible==2.10.7
    # via -r requirements.in
argcomplete==2.0.0
    # via yq
attrs==21.4.0
    # via jsonschema
bcrypt==3.2.0
    # via paramiko
borgbackup==1.1.17
    # via -r requirements.in
cached-property==1.5.2
    # via docker-compose
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
cryptography==36.0.2
    # via
    #   ansible-base
    #   paramiko
cython==0.29.21
    # via -r requirements.in
distro==1.7.0
    # via docker-compose
docker-compose==1.29.2
    # via -r requirements.in
docker[ssh]==5.0.3
    # via docker-compose
dockerpty==0.4.1
    # via docker-compose
docopt==0.6.2
    # via docker-compose
fabric==2.7.0
    # via -r requirements.in
humanize==2.4.0
    # via -r requirements.in
idna==3.3
    # via requests
importlib-metadata==4.8.3
    # via
    #   argcomplete
    #   jsonschema
invoke==1.7.1
    # via fabric
jinja2==3.0.3
    # via ansible-base
jsonschema==3.2.0
    # via docker-compose
markupsafe==2.0.1
    # via jinja2
packaging==21.3
    # via
    #   ansible-base
    #   borgbackup
paramiko==2.10.3
    # via
    #   docker
    #   fabric
pathlib2==2.3.7.post1
    # via fabric
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via paramiko
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.6.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via docker-compose
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   ansible-base
    #   docker-compose
    #   yq
requests==2.27.1
    # via
    #   -r requirements.in
    #   docker
    #   docker-compose
six==1.16.0
    # via
    #   bcrypt
    #   dockerpty
    #   jsonschema
    #   paramiko
    #   pathlib2
    #   python-dateutil
    #   websocket-client
texttable==1.6.4
    # via docker-compose
toml==0.10.2
    # via yq
typing-extensions==4.1.1
    # via importlib-metadata
urllib3==1.26.3
    # via
    #   -r requirements.in
    #   requests
websocket-client==0.59.0
    # via
    #   docker
    #   docker-compose
xmltodict==0.13.0
    # via yq
yq==3.1.0
    # via -r requirements.in
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
